ai-img-backend/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
diffusers==0.35.1
dotenv==0.9.9
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
Pillow==10.1.0
protobuf==6.32.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
openai==0.28
